handled errors and optimized middleware functions using controller

+++ backend/routes/posts.js
@@ -1,36 +1,38 @@
 const express = require("express");
+const multer = require("multer");
 
 const Post = require("../models/post");
-const multer = require("multer");
 const checkAuth = require("../middleware/check-auth");
+
 const router = express.Router();
 
 const MIME_TYPE_MAP = {
-  'image/png': 'png',
-  'image/jpeg': 'jpg',
-  'image/jpg': 'jpg'
+  "image/png": "png",
+  "image/jpeg": "jpg",
+  "image/jpg": "jpg"
 };
 
-// configure multer to enable route to be able to receive files
 const storage = multer.diskStorage({
-destination: (req, file, cb) => {
-const isValid = MIME_TYPE_MAP[file.mimetype];
-let error = new Error("Invalid mime type");
-if(isValid){
-  error = null;
-}  
-cb(error, "backend/images");
-},
-filename: (req, file, cb) => {
-  const name = file.originalname.toLocaleLowerCase().split(' ').join('-');
-  const ext = MIME_TYPE_MAP[file.mimetype];
-  cb(null, name + '-' + Date.now() + '.' + ext);
-}
+  destination: (req, file, cb) => {
+    const isValid = MIME_TYPE_MAP[file.mimetype];
+    let error = new Error("Invalid mime type");
+    if (isValid) {
+      error = null;
+    }
+    cb(error, "backend/images");
+  },
+  filename: (req, file, cb) => {
+    const name = file.originalname
+      .toLowerCase()
+      .split(" ")
+      .join("-");
+    const ext = MIME_TYPE_MAP[file.mimetype];
+    cb(null, name + "-" + Date.now() + "." + ext);
+  }
 });
 
 router.post(
   "",
-  // Restrict this route to only authenticated users
   checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
@@ -38,7 +40,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -48,55 +51,69 @@ router.post(
           id: createdPost._id
         }
       });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: 'Creating a post failed!'
+      });
     });
   }
 );
 
-router.put("/:id", 
-// Restrict this route to only authenticated users
-checkAuth,
-multer({storage: storage}).single("image"), (req, res, next) => {
-  let imagePath = req.body.imagePath;
-  if(req.file){
-    const url = req.protocol + "://" + req.get("host");
-    imagePath = url + "/images/" + req.file.filename
+router.put(
+  "/:id",
+  checkAuth,
+  multer({ storage: storage }).single("image"),
+  (req, res, next) => {
+    let imagePath = req.body.imagePath;
+    if (req.file) {
+      const url = req.protocol + "://" + req.get("host");
+      imagePath = url + "/images/" + req.file.filename;
+    }
+    const post = new Post({
+      _id: req.body.id,
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
+    });
+    console.log(post);
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" });
+      }
+    })
+    // catch is used when there is a technical error
+    .catch(error => {
+      res.status(500).json({
+      message: "Couldn't update post!"
+      });
+    });
   }
-  console.log(req.file);
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath
-  });
-  console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: "Update successful!" });
-  });
-});
+);
 
 router.get("", (req, res, next) => {
   const pageSize = +req.query.pagesize;
   const currentPage = +req.query.page;
   const postQuery = Post.find();
   let fetchedPosts;
-  if(pageSize && currentPage){
-    postQuery
-    .skip(pageSize * (currentPage - 1))
-    // Limit the number of items per page
-     .limit(pageSize);
+  if (pageSize && currentPage) {
+    postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
   }
-  postQuery.then(documents => {
-    fetchedPosts = documents;
-    return Post.count();  
+  postQuery
+    .then(documents => {
+      fetchedPosts = documents;
+      return Post.count();
     })
-    // Get the count of posts available
     .then(count => {
       res.status(200).json({
         message: "Posts fetched successfully!",
         posts: fetchedPosts,
         maxPosts: count
+      });
     });
-  });
 });
 
 router.get("/:id", (req, res, next) => {
@@ -106,16 +123,22 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({ message: "Post not found!" });
     }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching posts failed!"
+    });
   });
 });
 
-router.delete("/:id",
-// Restrict this route to only authenticated users
-checkAuth,
-(req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+router.delete("/:id", checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   });
 });
